Objects now wrap around within camera view
The bounds have an optional padding amount to avoid objects disappearing when they are still visible at the edge of the screen.

In a perfect world the padding amount would be dynamic and increase in proportion to the objects size so that objects of arbitrary sizes will never just "disappear" when they touch the edge of the screen before they are off the screen.

This could be tackled in a future commit when 2D Colliders are implemented, or by using render bounds, however, in my experience render bounds (at least in 3D, non DOTS unity) are heavily optimized for performance and not accuracy or consistency and so I would likely prefer using collisions > render bounds for these purposes.

- also debugged input asset: now using A/D instead of Q/E to turn.