Added masks so destructible's can decide what they're destroyed by
Explanation:
Now anything can destroy and/or be destroyed by other specified entities when they overlap.

Rational:
It's nice to have a generic system like this especially if everything is a one-hit-destruction.
If HP or damage based weapons / interactions are to be added, then this system will not support this at the moment.