Debugging wave spawner
Continuing work on wave spawner.  Using system state components to allow the spawner to track the lifetime of the enemies it spawns.

Problems:
There's lots of unreproducible bugs that have been occurring without changes to code or project (often just turning the computer on will trigger or resolve them). I think they're internal Unity bugs because of this, so I may upgrade Unity project versions to see if this fixes things.
- Using FixedLists that are frequently changed in the spawner is awkward, error prone (and likely slow) because they are all copied by value. Thinking I might want to use AssetBlobArrays here instead.

Other changes:
- Replaced most unsafe singleton calls like GetSingleton with the safe TryGetSingleton. There were a lot of issues with there being more than two singletons in the scene (the WorldTime kept duplicating in the editor), but I now assume this was an internal Unity bug and had little to do with my user code.

- Collider2D authoring now adapts to the localscale of the gameObject to improve convenience.

- Bullets now come out rotated with ship's direction, are tap-to-fire, and are destroyed on collisions with asteroids (instead of passing through them)

- Changed the rendering from sprite renderers to mesh renderers

- Added different asteroid sizes, Added more Waves