Asteroids now spawn smaller asteroids on destruction
Features:
Made destruction system more abstract by making it flag impending destruction, allowing other systems to respond to these "events" over the course of the frame in a decoupled manner.

Now there is a generic buffer & system to instantiate an arbitrary number of entities on the destruction of an entity.

Thoughts/Improvements:
Now that destruction is flagged we don't really still need the Spawned component to be a SystemComponent for the spawning systems to clean up and respond to their spawned entities being destroyed., instead it can just be a regular ComponentData.

Bugs:
The random direction that a newly instantiated entity does not work properly.

Other:
removed unused ILauncher class